centralize free tier limits
